+++ relm4-macros/src/widgets/gen/assign/signal_handler.rs
@@ -15,7 +15,7 @@ impl SignalHandler {
         sender_name: &Ident,
     ) {
         let span = p_name.span();
-        let assign_fn = p_name.assign_fn_stream(info.widget_name);
+        let assign_fn = p_name.assign_fn_stream(info);
         let self_assign_args = p_name.assign_args_stream(info.widget_name);
 
         let (clone_stream, assignment) = match &self.inner {