+++ relm4-macros/src/widgets/gen/util/property_name.rs
@@ -2,17 +2,28 @@ use proc_macro2::TokenStream as TokenStream2;
 use quote::{quote, quote_spanned, ToTokens};
 use syn::Ident;
 
-use crate::widgets::gen::PropertyName;
+use crate::widgets::gen::{assign::AssignInfo, PropertyName};
 
 impl PropertyName {
-    pub(crate) fn assign_fn_stream(&self, w_name: &Ident) -> TokenStream2 {
+    pub(crate) fn assign_fn_stream(&self, info: &mut AssignInfo<'_>) -> TokenStream2 {
+        let AssignInfo {
+            widget_name,
+            template_name,
+            ..
+        } = info;
+        let widget_name = if let Some(template_name) = template_name {
+            quote! { #template_name.#widget_name }
+        } else {
+            quote! { #widget_name }
+        };
+
         match self {
             PropertyName::Ident(ident) => {
-                quote! { #w_name.#ident }
+                quote! { #widget_name.#ident }
             }
             PropertyName::Path(path) => path.to_token_stream(),
             PropertyName::RelmContainerExtAssign(span) => {
-                quote_spanned! { *span => #w_name.container_add }
+                quote_spanned! { *span => #widget_name.container_add }
             }
         }
     }