+++ CHANGES.md
@@ -6,6 +6,7 @@
 
 + core: Fix an issue with using `connect_open` on `gtk::Application`
 + macros: Allow trailing commas in view!
++ macros: Allow multiple instances of the same template children
 + examples: Fix libadwaita tab examples
 
 ### Changed